# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Elaniin",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Log in End points"
                ],
                "summary": "Log in",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "example": "admin@admin.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345678"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Log in End points"
                ],
                "summary": "Register new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "example": "admin"
                    },
                    {
                        "name": "userName",
                        "in": "query",
                        "description": "UserName",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "example": "test@gmail.com"
                    },
                    {
                        "name": "birthDate",
                        "in": "query",
                        "description": "birthDate",
                        "example": "2062-02-26"
                    },
                    {
                        "name": "telephone",
                        "in": "query",
                        "description": "telephone",
                        "example": "77553344"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "example": "12345678"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "password_confirmation",
                        "required": true,
                        "example": "12345678"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data save successfully"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Oops something went wrong."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/logout/": {
            "post": {
                "tags": [
                    "Log in End points"
                ],
                "summary": "destroy token",
                "operationId": "0d1472f2c3c9109e6cbf60280872337a",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/refresh/": {
            "post": {
                "tags": [
                    "Log in End points"
                ],
                "summary": "refresh token",
                "operationId": "41fb937c56f3c3687c27626af72e69e5",
                "responses": {
                    "200": {
                        "description": "Return refresh token"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/user/": {
            "get": {
                "tags": [
                    "Log in End points"
                ],
                "summary": "return user authenticated ",
                "operationId": "9d4b56510a51413fdcba6ff44b97142b",
                "responses": {
                    "200": {
                        "description": "Return user authenticated"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}